pyproject.toml
[project]
name = "multi-modal-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.2",
    "jinja2>=3.1.6",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.21",
    "langchain-groq>=1.1.2",
    "pathlib>=1.0.1",
    "pillow>=12.1.1",
    "pypdf>=6.9.2",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.42.0",
]
